+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `ReadableStream` now extends `Stream` and can be used to `CopyTo` other .NET streams.
 
 ## [0.1.0] - 2022-10-30
 ### Added